packages/nobes-image/pyproject.toml
[project]
name = "nobes-image"
description = """Image manipulation, OCR, encoding"""
authors = [
  {name = "sneakers-the-rat", email = "sneakers-the-rat@protonmail.com"},
]
dynamic=["version"]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "EUPL-1.2"}
keywords = [
  "noob",
  "noob-nodes",
  "nobes",
]
dependencies = [
  "noob",
  "Pillow>=12.1.1",
  "pytesseract>=0.3.13",
]

[project.optional-dependencies]
tests = [
    "pytest",
]
dev = [
    "nobes-image[tests]",
]

[project.urls]
homepage = "https://noob.readthedocs.io"
documentation = "https://noob.readthedocs.io"
repository = "https://github.com/miniscope/noob"
changelog = "https://noob.readthedocs.io/en/latest/changelog.html"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.build]
includes = ["nobes"]

[tool.pdm.version]
# versions from tags like "v0.1.2"
fallback_version = "0.0.0"
source = "scm"
tag_filter="nobes-image-v*"
tag_regex = '^nobes-image-v(?:\D*)?(?P<version>([1-9][0-9]*!)?(0|[1-9][0-9]*)(\.(0|[1-9][0-9]*))*((a|b|c|rc)(0|[1-9][0-9]*))?(\.post(0|[1-9][0-9]*))?(\.dev(0|[1-9][0-9]*))?$)$'
